requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
joblib==1.4.2
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12
